+++ src/components/Layout/style.css
@@ -14,7 +14,7 @@
         position: absolute;
         bottom: -200px;
         left: 0;
-        background: url(../image/bg3.jpg) bottom left repeat-x;
+        background: url(../../image/bg3.jpg) bottom left repeat-x;
         background-size: cover;
         width: 100%;
         height: 100%;
@@ -25,7 +25,7 @@
         position: absolute;
         bottom: -300px;
         left: 0;
-        background: url(../image/bg1.jpg) bottom left repeat-x;
+        background: url(../../image/bg1.jpg) bottom left repeat-x;
         background-size: cover;
         width: 100%;
         height: 100%;
